# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: 智能用药App API
  description: 智能用药App后端API接口文档 - 支持慢性病患者和老年人的用药管理
  version: "1.0.0"
  contact:
    name: 后端开发团队
    email: backend@smartmed.com
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://api.smartmed.com/api
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用JWT Bearer Token进行认证，通过登录接口获取
  schemas:
...
    PrescriptionMedicine:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
    CreatePrescriptionRequest:
      type: object
      properties:
        doctorName:
          type: string
        doctorTitle:
          type: string
        hospitalName:
          type: string
        department:
          type: string
        prescriptionNumber:
          type: string
        issueDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        diagnosis:
          type: string
        medicines:
          type: array
          items:
            $ref: '#/components/schemas/PrescriptionMedicine'
        imageUrl:
          type: string
          format: uri
        notes:
          type: string
...
paths:
...
  /prescriptions:
    get:
      tags: [处方管理]
      summary: 获取处方列表
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
    post:
      tags: [处方管理]
      summary: 创建处方
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrescriptionRequest'
      responses:
        '201':
          description: 处方已创建
  /prescriptions/{id}:
    get:
      tags: [处方管理]
      summary: 获取处方详情
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
    put:
      tags: [处方管理]
      summary: 更新处方
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 处方已更新
    delete:
      tags: [处方管理]
      summary: 删除处方
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 处方已删除
  /prescriptions/{id}/ocr:
    post:
      tags: [处方管理]
      summary: 提交OCR识别
      description: 上传处方图片进行OCR识别
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - imageUrl
              properties:
                imageUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: OCR识别已提交

  # ========== Guardians ==========
...